# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HMS Elite API
  version: 1.0.0
  description: API v1 para HMS Elite (auth, rooms, bookings, guests, users).
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - cookieAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/refresh:
    post:
      summary: Refresh token
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refresh exitoso
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/auth/logout:
    post:
      summary: Logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Logout exitoso
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  responses:
    ErrorResponse:
      description: Error estándar
      headers:
        x-api-version:
          schema:
            type: string
          description: Versión de contrato de API.
        x-api-deprecation-policy:
          schema:
            type: string
          description: Política de compatibilidad/deprecación del contrato.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
    LoginRequest:
      type: object
      required: [hotel_id, username, password]
      properties:
        hotel_id:
          type: string
        username:
          type: string
        password:
          type: string
        device_id:
          type: string
          description: Identificador estable del dispositivo/sesión cliente.
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
        all_devices:
          type: boolean
          description: Si es true en logout, revoca sesiones de todos los dispositivos.
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        role:
          type: string
...